requirements.txt
flask
tensorflow
opencv-python-headless
numpy
Pillow